requirements.txt
annotated-types==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
anyio==4.3.0 ; python_version >= "3.10" and python_version < "4.0"
botocore==1.34.91 ; python_version >= "3.10" and python_version < "4.0"
build==1.2.1 ; python_version >= "3.10" and python_version < "4.0"
cachecontrol[filecache]==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.3.3 ; python_version >= "3.10" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.10" and python_version < "4.0" and (sys_platform == "darwin" or sys_platform == "linux") and (sys_platform == "darwin" or platform_python_implementation != "PyPy")
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0"
cleo==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and (platform_system == "Windows" or os_name == "nt")
crashtest==0.4.1 ; python_version >= "3.10" and python_version < "4.0"
cryptography==42.0.5 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "linux"
distlib==0.3.8 ; python_version >= "3.10" and python_version < "4.0"
dulwich==0.21.7 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.1 ; python_version >= "3.10" and python_version < "3.11"
fastapi==0.110.2 ; python_version >= "3.10" and python_version < "4.0"
fastjsonschema==2.19.1 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.13.4 ; python_version >= "3.10" and python_version < "4.0"
google-api-core==2.18.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.29.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-storage==2.16.0 ; python_version >= "3.10" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
google-resumable-media==2.7.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.63.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.7 ; python_version >= "3.10" and python_version < "4.0"
importlib-metadata==7.1.0 ; python_version >= "3.10" and python_version < "3.12"
installer==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
jaraco-classes==3.4.0 ; python_version >= "3.10" and python_version < "4.0"
jeepney==0.8.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "linux"
jinja2==3.1.3 ; python_version >= "3.10" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
keyring==24.3.1 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.10" and python_version < "4.0"
more-itertools==10.2.0 ; python_version >= "3.10" and python_version < "4.0"
msgpack==1.0.8 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.10" and python_version < "4.0"
pastel==0.2.1 ; python_version >= "3.10" and python_version < "4.0"
pexpect==4.9.0 ; python_version >= "3.10" and python_version < "4.0"
pkginfo==1.10.0 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==4.2.1 ; python_version >= "3.10" and python_version < "4.0"
poethepoet==0.25.1 ; python_version >= "3.10" and python_version < "4.0"
poetry-core==1.9.0 ; python_version >= "3.10" and python_version < "4.0"
poetry-plugin-export==1.7.1 ; python_version >= "3.10" and python_version < "4.0"
poetry==1.8.2 ; python_version >= "3.10" and python_version < "4.0"
proto-plus==1.23.0 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.25.3 ; python_version >= "3.10" and python_version < "4.0"
ptyprocess==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.10" and python_version < "4.0" and (sys_platform == "darwin" or sys_platform == "linux") and (sys_platform == "darwin" or platform_python_implementation != "PyPy")
pydantic-core==2.18.2 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.7.1 ; python_version >= "3.10" and python_version < "4.0"
pygments==2.17.2 ; python_version >= "3.10" and python_version < "4.0"
pyproject-hooks==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
python-multipart==0.0.9 ; python_version >= "3.10" and python_version < "4.0"
pywin32-ctypes==0.2.2 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
rapidfuzz==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
replicate==0.25.2 ; python_version >= "3.10" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
s3transfer==0.10.1 ; python_version >= "3.10" and python_version < "4.0"
secretstorage==3.3.3 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "linux"
shellingham==1.5.4 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.37.2 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
tomlkit==0.12.4 ; python_version >= "3.10" and python_version < "4.0"
trove-classifiers==2024.4.10 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.11.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.29.0 ; python_version >= "3.10" and python_version < "4.0"
virtualenv==20.26.0 ; python_version >= "3.10" and python_version < "4.0"
xattr==1.1.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "darwin"
zipp==3.18.1 ; python_version >= "3.10" and python_version < "3.12"
